backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.0

# Document processing
PyMuPDF==1.23.8
python-docx==1.1.0
pdfplumber==0.10.3

# Vector store
qdrant-client==1.7.0

# AI/ML
google-generativeai==0.3.1
sentence-transformers==2.2.2
numpy==1.26.2

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
